requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
discord.py==2.5.2
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.5.0
propcache==0.3.2
pyee==13.0.0
python-dotenv==1.1.0
requests==2.32.4
sniffio==1.3.1
soupsieve==2.7
typing_extensions==4.14.0
urllib3==2.5.0
yarl==1.20.1